Rename file to match contents
