+++ static/app/views/dashboards/utils/transformEventsStatsToSeries.tsx
